Added timing interval extension
Couldn't re-use https://simplifier.net/ColumnaCura/timing-interval/rendered as it's in draft and Systematic did not respond to a request for collaboration.

Modelled to be in line with that extension however for maximum compability. Only the context was changed to Timing.repeat from Timing itself, as the interval is only involved when the events are repeating.

Note: when this extension is used to represent Benevning tidsenhet (OID=9088), the week must be constrained out.