Renamed free text to coded text
"Free-text dosage timing" when it was really codes made no sense, whereas coded dosage timing makes a bit more, since it is coded.

It obviously raises the question as to why isn't MedicationStatement.dosage.timing isn't used for this - since that is a structured way to represent timing information - so I'll add a note into the profile to mention that this is for legacy reasons.